requirements/base.txt
Django>=5.0,<7.0
django-environ>=0.11
Pillow>=10.0
django-filter>=24.0
django-import-export>=4.0
django-simple-history>=3.5
whitenoise>=6.5
reportlab>=4.0
anthropic>=0.40
djangorestframework>=3.15
django-cors-headers>=4.3
requests>=2.31
django-redis>=5.4
sentry-sdk[django]>=2.0
openpyxl>=3.1
python-barcode>=0.15
qrcode[pil]>=7.4
channels>=4.0
channels-redis>=4.2
daphne>=4.0
celery[redis]>=5.3
django-celery-beat>=2.6
djangorestframework-simplejwt>=5.3
django-prometheus>=2.3
